Fix migrations, add test to vehicle model

+++ db/migrate/20200314205358_create_waypoints.rb
@@ -1,8 +1,8 @@
 class CreateWaypoints < ActiveRecord::Migration[5.2]
   def change
     create_table :waypoints do |t|
-      t.decimal :latitude, precision: 3, scale: 16
-      t.decimal :longitude, precision: 3, scale: 16
+      t.decimal :latitude, precision: 16, scale: 16
+      t.decimal :longitude, precision: 16, scale: 16
       t.datetime :send_at
       t.references :vehicle, foreign_key: true
